Update:prevent unauthorized access

+++ app/Controllers/Http/ThreadsController.ts
@@ -12,7 +12,7 @@ export default class ThreadsController {
     const categoryId = request.input('category_id')
 
     const payload = await request.validate(SortValidator)
-    // 
+
     const sortBy = payload.sort_by || 'id'
     const sortOrder = payload.sort_order || 'asc'
 
@@ -49,10 +49,14 @@ export default class ThreadsController {
     return response.ok(thread)
   }
 
-  public async update ({ params, request, response }: HttpContextContract) {
+  public async update ({ params, request, response, auth }: HttpContextContract) {
     const thread = await Thread.findOrFail(params.id)
 
     const payload = await request.validate(UpdateThreadValidator)
+
+    if(auth.user?.id !== thread.userId){
+      return response.unauthorized({message: 'You are not authorized to update this thread.'})
+    }
 
     thread.merge(payload)
     await thread.save()
@@ -64,8 +68,12 @@ export default class ThreadsController {
     return response.ok(thread)
   }
 
-  public async destroy ({ params }: HttpContextContract) {
+  public async destroy ({ params, auth, response }: HttpContextContract) {
     const thread = await Thread.findOrFail(params.id)
+
+    if(auth.user?.id !== thread.userId){
+      return response.unauthorized({message: 'You are not authorized to delete this thread.'})
+    }
 
     await thread.delete()
   }
